Move all tests to their own directories
