+++ app/components/Annotation/__tests__/Annotation-test.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { expect } from 'chai';
 import Immutable from 'immutable';
-import { defaultLabels } from '../../defaults';
+import { defaultLabels } from '../../../defaults';
 import sinon from 'sinon';
 
 // see: https://github.com/mochajs/mocha/issues/1847
 const { describe, it } = global;
 
-import Annotation from '../Annotation/presenter';
+import Annotation from '../presenter';
 
 describe('<Annotation />', () => {
